# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API",
        "description": "API documentation for the Library application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List all books",
                "description": "Returns a list of books",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Creates a new book",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a book by ID",
                "description": "Returns a single book",
                "operationId": "e94109dcb4f4d40b58a210b43d94016d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Update an existing book",
                "description": "Updates an existing book",
                "operationId": "7430c1078c3cb5da552b2fd50409f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Book"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Delete a book",
                "description": "Deletes a book by ID",
                "operationId": "41c44a46960ad1a921ba5bfbf16ea7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Book deleted"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the author",
                        "type": "string"
                    },
                    "birth_date": {
                        "description": "Birth date of the author",
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "description": "Timestamp when the author was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the author was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Book": {
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "publication_year": {
                        "description": "Publication year of the book",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "description": "Timestamp when the book was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the book was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "authors": {
                        "description": "List of authors associated with the book",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}